fix(Stat): Search stat
