+++ src/middlewares/index.js
@@ -40,7 +40,6 @@ export const jwtParseMiddleware = (req, res, next) => {
 
 export const validateId = (key) => (req, res, next) => {
   const value = req[key] || req.params[key] || req.body[key] || req.query[key];
-  console.log (req.params[key]);
   if (!value) {
     logger.api.error (`[${req.method}] ${req.originalUrl} request error; 400 - empty ${key}`);
     return res.status (400).json ({
