+++ src/utils/statistic.js
@@ -1,27 +1,12 @@
-import { Statistic, User } from '../db';
-import { logger } from './logger';
-
-
+import { Statistic } from '../db';
 import { EVENTS, EVENTS_MAP } from './variables';
-
-const stat = EVENTS.reduce ((acc, cur) => ({
-  ...acc,
-  [cur]: (data) => Statistic.create ({
-    type: cur,
-    data,
-  })
-    .catch (error => {
-      logger.event.error (`Creating ${cur} Stat Failed`);
-      logger.event.error (error);
-    }),
-}), {});
 
 export const statZabo = async ({ zaboId, decoded }) => {
   if (!decoded) return null;
   const data = {
     type: EVENTS_MAP.GET_ZABO,
-    zaboId,
-    sso_sid: decoded.sid,
+    zabo: zaboId,
+    user: decoded._id || decoded.id,
   };
   return Statistic.create (data);
 };
@@ -33,9 +18,7 @@ export const statSearch = async ({ query, category, decoded }) => {
     category,
   };
   if (decoded) {
-    data.sso_sid = decoded.sid;
+    data.user = decoded._id || decoded.id;
   }
   return Statistic.create (data);
 };
-
-export { stat };
